03-language-service/requirements.txt
azure-ai-textanalytics==5.3.0
python-dotenv